backend/requirements.txt
astroid==2.3.2
autopep8==1.4.4
cffi==1.13.0
cryptography==2.8
Django==2.2.6
django-cors-headers==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysqlclient==1.4.4
pep8==1.7.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
